Move ddmin and mutators to separat files, smaller refactoring.
