+++ utils/mutators.py
@@ -1,0 +1,8 @@
+from utils import smtlib
+
+class PassDeleteAssert:
+    def filter(self, x):
+        return smtlib.is_assert(x)
+
+    def subst(self, x):
+        return None
